docs(menu): Change ControlButton to Basic example and update docs
